+++ lua/cbits/hslua/hslua.h
@@ -6,6 +6,8 @@ int hslua_call_hs(lua_State *L);
 
 int hslua_userdata_gc(lua_State *L);
 
+void hslua_arith(lua_State *L, int op, int *status);
+
 int hslua_compare(lua_State *L, int index1, int index2, int op, int *status);
 
 void hslua_concat(lua_State *L, int n, int *status);