+++ lua/test/Lua/ErsatzTests.hs
@@ -11,7 +11,9 @@ module Lua.ErsatzTests (tests) where
 
 import Control.Monad (void)
 import Foreign.C.String (withCStringLen)
+import Foreign.Marshal (alloca)
 import Foreign.Ptr (nullPtr)
+import Foreign.Storable (peek)
 import Lua
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit (Assertion, HasCallStack, testCase, (@=?) )
@@ -19,7 +21,68 @@ import Test.Tasty.HUnit (Assertion, HasCallStack, testCase, (@=?) )
 -- | Tests for unsafe methods.
 tests :: TestTree
 tests = testGroup "ersatz"
-  [ testGroup "global"
+  [ testGroup "arith"
+    [ "adds two numbers" =: do
+        7 `shouldBeResultOf` \l -> do
+          lua_pushinteger l 5
+          lua_pushinteger l 2
+          hslua_arith l LUA_OPADD nullPtr
+          lua_tointegerx l top nullPtr
+    , "negates number" =: do
+        (-5) `shouldBeResultOf` \l -> do
+          lua_pushinteger l 5
+          hslua_arith l LUA_OPUNM nullPtr
+          lua_tointegerx l top nullPtr
+    , "pops its arguments from the stack" =: do
+        1 `shouldBeResultOf` \l -> do
+          old <- lua_gettop l
+          lua_pushinteger l 4
+          lua_pushinteger l 3
+          hslua_arith l LUA_OPADD nullPtr
+          new <- lua_gettop l
+          return (new - old)
+    , "pops a single argument for unary negation" =: do
+        1 `shouldBeResultOf` \l -> do
+          old <- lua_gettop l
+          lua_pushinteger l 3
+          hslua_arith l LUA_OPUNM nullPtr
+          new <- lua_gettop l
+          return (new - old)
+    , "sets status to LUA_OK on success" =: do
+        (LUA_OK, 1024) `shouldBeResultOf` \l -> do
+          lua_pushinteger l 2
+          lua_pushinteger l 10
+          stts <- alloca $ \status -> do
+            hslua_arith l LUA_OPPOW status
+            peek status
+          result <- lua_tointegerx l top nullPtr
+          return (stts, result)
+    , "runs metamethod" =: do
+        (LUA_OK, 7) `shouldBeResultOf` \l -> do
+          lua_pushinteger l 2
+          lua_createtable l 0 0
+          LUA_OK <- withCStringLen "return {__mod = function() return 7 end}" $
+            \(s, len) -> luaL_loadbuffer l s (fromIntegral len) s
+          LUA_OK <- lua_pcall l 0 1 0
+          lua_setmetatable l (-2)
+          stts <- alloca $ \status -> do
+            hslua_arith l LUA_OPMOD status
+            peek status
+          result <- lua_tointegerx l top nullPtr
+          return (stts, result)
+    , "catches runtime error in metamethod" =: do
+        LUA_ERRRUN `shouldBeResultOf` \l -> do
+          lua_pushinteger l 2
+          lua_createtable l 0 0
+          LUA_OK <- withCStringLen "return {__bor = function() error'no' end}" $
+            \(s, len) -> luaL_loadbuffer l s (fromIntegral len) s
+          LUA_OK <- lua_pcall l 0 1 0
+          lua_setmetatable l (-2)
+          alloca $ \status -> do
+            hslua_arith l LUA_OPBOR status
+            peek status
+    ]
+  , testGroup "global"
     [ "set and get global field" =: do
         (42, LUA_TNUMBER) `shouldBeResultOf` \l -> do
           lua_pushinteger l 42