+++ lua/CHANGELOG.md
@@ -28,6 +28,9 @@ Release pending.
     -   LUA_OPSHL
     -   LUA_OPSHR
 
+-   Added an ersatz function `hslua_arith` that wraps `lua_arith`
+    and catches errors that occur in metamethods.
+
 ## lua 2.0.2
 
 Released 2021-11-26.